# openapi.yaml
openapi: 3.0.3
info:
  title: Inventory Tracker API
  version: "1.0.0"
  description: >
    Manages 5 fixed products (apple, banana, grape, orange, watermelon):
    stock levels, sell price, purchase cost (매입가), selling, and net worth
    (sales revenue minus restock purchase costs).
servers:
  - url: http://127.0.0.1:5001
    description: Local development server

paths:
  /api/docs:
    get:
      summary: Download the API reference as a PDF
      responses:
        "200":
          description: The PDF file
          content:
            application/pdf: {}

  /api/products:
    get:
      summary: List all products
      responses:
        "200":
          description: All 5 products with current stock, sell price, and cost
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /api/products/{id}:
    get:
      summary: Look up a single product by id
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: The product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/products/{id}/restock:
    post:
      summary: Add 1 to stock_qty and log a purchase at current cost
      description: >
        Increases stock by 1 and records a purchase row so that cost is
        subtracted from net worth.
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: The updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/products/{id}/price:
    put:
      summary: Set the product's sell price
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [price]
              properties:
                price:
                  type: number
                  minimum: 0
                  example: 1.5
      responses:
        "200":
          description: The updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Missing or invalid price
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/products/{id}/cost:
    put:
      summary: Set the product's purchase cost (매입가)
      parameters:
        - $ref: "#/components/parameters/ProductId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cost]
              properties:
                cost:
                  type: number
                  minimum: 0
                  example: 0.8
      responses:
        "200":
          description: The updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Missing or invalid cost
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/products/{id}/sell:
    post:
      summary: Subtract 1 from stock_qty and log the sale
      parameters:
        - $ref: "#/components/parameters/ProductId"
      responses:
        "200":
          description: The updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          description: Out of stock
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/sales:
    get:
      summary: Transaction history (filter / sort / group by product)
      description: >
        Returns sales newest-first by default. Use query params to filter
        by product, sort by product name, or group into product categories.
      parameters:
        - name: product_id
          in: query
          required: false
          description: Only sales for this product id
          schema:
            type: integer
            example: 2
        - name: product
          in: query
          required: false
          description: Only sales for this product name
          schema:
            type: string
            example: banana
        - name: sort
          in: query
          required: false
          description: Order of the flat list (`sold_at` default, or `product`)
          schema:
            type: string
            enum: [sold_at, product]
            default: sold_at
        - name: group_by
          in: query
          required: false
          description: >
            If `product`, return an object keyed by product name
            (each value is that product's sales list) instead of a flat array
          schema:
            type: string
            enum: [product]
      responses:
        "200":
          description: >
            Flat list of sales, or (when group_by=product) an object of
            product_name → sales[]
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Sale"
                  - type: object
                    additionalProperties:
                      type: array
                      items:
                        $ref: "#/components/schemas/Sale"
        "400":
          description: Invalid query parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /api/valuation:
    get:
      summary: Net worth (sales revenue minus restock purchase costs)
      responses:
        "200":
          description: Net worth breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  net_worth:
                    type: number
                    example: 10.5
                  revenue:
                    type: number
                    example: 22.3
                  costs:
                    type: number
                    example: 11.8

components:
  parameters:
    ProductId:
      name: id
      in: path
      required: true
      description: Product id (1-5)
      schema:
        type: integer

  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: apple
        stock_qty:
          type: integer
          example: 3
        price:
          type: number
          description: Sell price
          example: 1.5
        cost:
          type: number
          description: Purchase cost (매입가)
          example: 0.8

    Sale:
      type: object
      properties:
        id:
          type: integer
        product_id:
          type: integer
          example: 2
        product_name:
          type: string
          example: banana
        qty_sold:
          type: integer
        price_at_sale:
          type: number
        sold_at:
          type: string
          format: date-time

    Error:
      type: object
      properties:
        error:
          type: string
          example: no product with that id

  responses:
    NotFound:
      description: No product with that id
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"